Separate Training and Analysis workspaces
Move observed training metrics out of the managed-plan workspace so Training stays focused on planning while Analysis owns diagnostics and activity history. Rework the five-slot miniapp navigation around Analysis and Me without breaking legacy heat-adaptation deep links.
