+++ .github/CODEOWNERS
@@ -40,12 +40,18 @@
 /web/src/pages/Goal.tsx             @dddtc2005
 /web/src/pages/Today.tsx            @dddtc2005
 /web/src/pages/Training.tsx         @dddtc2005
+/web/src/pages/Analysis.tsx         @dddtc2005
 /web/src/types/api.ts               @dddtc2005
 /miniapp/pages/science/             @dddtc2005
 /miniapp/pages/training/            @dddtc2005
+/miniapp/pages/analysis/            @dddtc2005
+/miniapp/pages/history/             @dddtc2005
+/miniapp/pages/me/                  @dddtc2005
 /miniapp/pages/today/               @dddtc2005
 /miniapp/pages/goal/                @dddtc2005
 /miniapp/pages/settings/            @dddtc2005
+/miniapp/components/activity-history/ @dddtc2005
+/miniapp/components/observed-training-switch/ @dddtc2005
 /miniapp/utils/heat-adaptation.ts   @dddtc2005
 /miniapp/utils/i18n-extra.ts        @dddtc2005
 /miniapp/types/api.ts               @dddtc2005
