+++ CLAUDE.md
@@ -29,7 +29,7 @@ Garmin/Stryd/Oura APIs → sync/*.py → db/sync_writer.py → SQLite (trainsigh
 | `api/` | FastAPI backend | `main.py`, `deps.py` (data layer — `get_dashboard_data()`), `auth.py` (JWT), `users.py`, `views.py`, `invitations.py`, `routes/` (incl. `admin.py`, `register.py`, `wechat.py`) |
 | `frontend_server/` | Static SPA host on its own App Service site (`praxys-frontend`) | `main.py` (`create_app(dist_dir)` factory, `SPAStaticFiles` with 404→`index.html` fallback for non-asset paths, cache-control middleware, `/healthz`). Decoupled from `api/` so the same `web/dist/` artifact can later sit on Tencent COS (CN audience, post-ICP) without Azure-specific glue. |
 | `web/src/` | React SPA | `pages/` (Today, Training, Goal, History, Science, Settings, Admin, Login, Setup), `components/`, `hooks/`, `contexts/`, `types/api.ts`, `lib/` |
-| `miniapp/` | WeChat Mini Program (native + Skyline + TypeScript) | `pages/` (login WebView, plus Skyline pages: today, training, goal, history, settings, science, legal), `components/` (`nav-bar` custom Skyline header, `line-chart` Canvas 2D), `utils/` (`api-client.ts` wx.request + JWT, `auth.ts` wx.login flow, `format.ts`/`share.ts`/`theme.ts`), `types/api.ts` auto-synced from `web/src/types/api.ts` by `scripts/sync-types.cjs`, and the EULA + Privacy text (`utils/legal.ts`) from `web/src/lib/legal.ts` by `scripts/sync-legal.cjs` (both run on `npm run typecheck`). The `legal` page renders those Terms/Privacy docs in-app and the login page links to them + gates the waitlist / account-link email capture behind an explicit consent checkbox (WeChat audit requirement). Build: WeChat DevTools handles TypeScript + Sass via `project.config.json` `useCompilerPlugins` — no webpack/babel toolchain. Open `miniapp/` as the DevTools project root; only devDeps are `miniprogram-api-typings` + `typescript` for `tsc --noEmit` CI checks. **Mini program position**: a view + manage companion to the web app. The mini program **owns the waitlist signup** as a CN-friendly entry point — WeChat is the primary discovery channel for that audience and the waitlist is just intent capture (email + optional note), no platform setup. **Full registration + the platform-connection wizard stays on praxys.run** because Garmin / Stryd / Oura OAuth is painful inside a miniapp WebView. Existing users link their account by email / password and then use the mini program day-to-day (signal, training, goal, sync, theory, training-base, theme / language). New users either join the waitlist inside WeChat or, if they already have an invitation code, are deep-linked to praxys.run to register. |
+| `miniapp/` | WeChat Mini Program (native + Skyline + TypeScript) | `pages/` (login WebView, plus Skyline pages: today, training, analysis, goal, me, history, settings, science, labs, legal), `components/` (`nav-bar` custom Skyline header, `line-chart` Canvas 2D), `utils/` (`api-client.ts` wx.request + JWT, `auth.ts` wx.login flow, `format.ts`/`share.ts`/`theme.ts`), `types/api.ts` auto-synced from `web/src/types/api.ts` by `scripts/sync-types.cjs`, and the EULA + Privacy text (`utils/legal.ts`) from `web/src/lib/legal.ts` by `scripts/sync-legal.cjs` (both run on `npm run typecheck`). The five primary tabs are Today, Training, Analysis, Goal, and Me; Activities lives inside Analysis, while Settings, Science, Labs, and legal surfaces live under Me. The `legal` page renders those Terms/Privacy docs in-app and the login page links to them + gates the waitlist / account-link email capture behind an explicit consent checkbox (WeChat audit requirement). Build: WeChat DevTools handles TypeScript + Sass via `project.config.json` `useCompilerPlugins` — no webpack/babel toolchain. Open `miniapp/` as the DevTools project root; only devDeps are `miniprogram-api-typings` + `typescript` for `tsc --noEmit` CI checks. **Mini program position**: a view + manage companion to the web app. The mini program **owns the waitlist signup** as a CN-friendly entry point — WeChat is the primary discovery channel for that audience and the waitlist is just intent capture (email + optional note), no platform setup. **Full registration + the platform-connection wizard stays on praxys.run** because Garmin / Stryd / Oura OAuth is painful inside a miniapp WebView. Existing users link their account by email / password and then use the mini program day-to-day (signal, managed training, analysis and activities, goal, sync, theory, training-base, theme / language). New users either join the waitlist inside WeChat or, if they already have an invitation code, are deep-linked to praxys.run to register. |
 | `plugins/praxys/` | Praxys plugin (**git submodule** of public [`praxys-run/praxys-coach-plugin`](https://github.com/praxys-run/praxys-coach-plugin), MIT) | `skills/` (8 SKILL.md), `mcp-server/` (local + remote MCP). Edits to anything inside `plugins/praxys/` are commits in the plugin repo, not this one — PRs there, then bump the submodule pointer here. |
 | `tests/` | pytest suite | |
 | `data/` | Fixtures + science YAML | `sample/` (test CSVs — not live data), `science/` (theory YAMLs) |
