+++ miniapp/components/activity-history/index.json
@@ -1,0 +1,6 @@
+{
+  "component": true,
+  "usingComponents": {
+    "state-card": "/components/state-card/index"
+  }
+}
