+++ miniapp/components/observed-training-switch/index.json
@@ -1,0 +1,3 @@
+{
+  "component": true
+}
