+++ miniapp/components/observed-training-switch/index.wxml
@@ -1,0 +1,27 @@
+<view
+  class="observed-switch"
+  role="tablist"
+  aria-label="{{tr.label}}"
+  aria-hidden="{{disabled}}"
+>
+  <view
+    class="observed-switch__item {{active === 'analysis' ? 'observed-switch__item--active' : ''}} {{disabled ? 'observed-switch__item--disabled' : ''}}"
+    data-value="analysis"
+    bindtap="onPick"
+    role="tab"
+    aria-selected="{{active === 'analysis'}}"
+    aria-disabled="{{disabled}}"
+  >
+    <text>{{tr.analysis}}</text>
+  </view>
+  <view
+    class="observed-switch__item {{active === 'activities' ? 'observed-switch__item--active' : ''}} {{disabled ? 'observed-switch__item--disabled' : ''}}"
+    data-value="activities"
+    bindtap="onPick"
+    role="tab"
+    aria-selected="{{active === 'activities'}}"
+    aria-disabled="{{disabled}}"
+  >
+    <text>{{tr.activities}}</text>
+  </view>
+</view>
