+++ miniapp/pages/analysis/index.json
@@ -1,0 +1,16 @@
+{
+  "orientation": "portrait",
+  "usingComponents": {
+    "nav-bar": "/components/nav-bar/index",
+    "line-chart": "/components/line-chart/index",
+    "scatter-chart": "/components/scatter-chart/index",
+    "bar-chart": "/components/bar-chart/index",
+    "state-card": "../../components/state-card/index",
+    "insight-feedback": "/components/insight-feedback/index",
+    "observed-training-switch": "/components/observed-training-switch/index",
+    "activity-history": "/components/activity-history/index"
+  },
+  "backgroundColor": "@bgColor",
+  "backgroundColorBottom": "@bgColor",
+  "backgroundColorContent": "@bgColor"
+}
