+++ miniapp/pages/history/index.wxml
@@ -1,5 +1,10 @@
-<view class="history-root {{themeClass}}">
-  <nav-bar title="{{tr.navTitle}}" theme-class="{{themeClass}}" />
+<view class="history-root {{themeClass}} {{languageClass}}">
+  <nav-bar
+    title="{{navTitle}}"
+    theme-class="{{themeClass}}"
+    show-back="{{true}}"
+    bindback="onBack"
+  />
 
   <scroll-view
     scroll-y="{{true}}"
@@ -7,79 +12,11 @@
     refresher-triggered="{{refreshing}}"
     bindrefresherrefresh="onScrollRefresh"
     bindscrolltolower="onScrollToBottom"
-    lower-threshold="{{200}}"
+    lower-threshold="180"
     class="history-scroll"
   >
-    <view class="history-content">
-      <text wx:if="{{hasActivities}}" class="history-total ts-muted">{{totalLine}}</text>
-
-      <block wx:if="{{loading && !hasActivities}}">
-        <state-card type="loading" />
-        <state-card type="loading" />
-        <state-card type="loading" />
-      </block>
-
-      <block wx:elif="{{errorMessage && !hasActivities}}">
-        <state-card type="error" headline="{{tr.failedToLoad}}" detail="{{errorMessage}}" show-retry bindretry="onRetry" />
-      </block>
-
-      <block wx:else>
-        <view
-          wx:for="{{activities}}"
-          wx:key="id"
-          class="ts-card history-row"
-          data-id="{{item.id}}"
-          bindtap="toggleExpand"
-        >
-          <view class="history-row-head">
-            <!-- Left-side disclosure chevron only when splits are present.
-                 Placed before date so it reads as "this row expands", not
-                 "this type expands". Rotation indicates expanded state. -->
-            <text
-              wx:if="{{item.hasSplits}}"
-              class="history-row-chev ts-muted {{item.expanded ? 'history-row-chev--open' : ''}}"
-            >›</text>
-            <text class="history-row-date ts-muted {{item.hasSplits ? 'history-row-date--indent' : ''}}">{{item.date}}</text>
-            <text class="history-row-type">{{item.type}}</text>
-          </view>
-          <view class="history-row-metrics">
-            <view
-              wx:for="{{item.metrics}}"
-              wx:key="label"
-              wx:for-item="metric"
-              class="history-metric"
-            >
-              <text class="history-metric-label ts-muted">{{metric.label}}</text>
-              <text class="history-metric-value ts-value">{{metric.value}}</text>
-            </view>
-          </view>
-
-          <view wx:if="{{item.hasSplits && item.expanded}}" class="history-splits">
-            <text class="ts-section-label">{{tr.splits}} ({{item.splitCount}})</text>
-            <view
-              wx:for="{{item.splitsDisplay}}"
-              wx:key="num"
-              wx:for-item="split"
-              class="history-split-row"
-            >
-              <text class="history-split-num ts-muted">{{split.num}}</text>
-              <text
-                wx:for="{{split.cells}}"
-                wx:key="*this"
-                wx:for-item="cell"
-                class="history-split-cell ts-value"
-              >{{cell}}</text>
-            </view>
-            <text wx:if="{{item.hasMoreSplits}}" class="ts-muted history-split-more">
-              + {{item.moreSplitsCount}} {{tr.more}}
-            </text>
-          </view>
-
-        </view>
-
-        <text wx:if="{{loadingMore}}" class="history-footer ts-muted">{{tr.loadingMore}}</text>
-        <text wx:elif="{{hasReachedEnd}}" class="history-footer ts-muted">{{tr.endOfActivities}}</text>
-      </block>
-    </view>
+    <activity-history
+      id="legacy-activity-history"
+    />
   </scroll-view>
 </view>
