+++ miniapp/pages/me/index.json
@@ -1,8 +1,7 @@
 {
   "orientation": "portrait",
   "usingComponents": {
-    "nav-bar": "/components/nav-bar/index",
-    "activity-history": "/components/activity-history/index"
+    "nav-bar": "/components/nav-bar/index"
   },
   "backgroundColor": "@bgColor",
   "backgroundColorBottom": "@bgColor",
