+++ miniapp/pages/settings/index.scss
@@ -74,31 +74,6 @@
   margin-bottom: 16rpx;
 }
 
-.settings-nav-row {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.settings-nav-title {
-  display: block;
-  font-size: 28rpx;
-  font-weight: 600;
-  color: var(--text);
-  margin-bottom: 4rpx;
-}
-
-.settings-nav-sub {
-  display: block;
-  font-size: 22rpx;
-  line-height: 1.4;
-}
-
-.settings-nav-chevron {
-  font-size: 36rpx;
-  padding-left: 24rpx;
-}
-
 .settings-origin {
   display: block;
   font-size: 18rpx;
