+++ miniapp/pages/settings/index.wxml
@@ -1,5 +1,10 @@
 <view class="settings-root {{themeClass}}">
-  <nav-bar title="{{tr.navTitle}}" theme-class="{{themeClass}}" />
+  <nav-bar
+    title="{{tr.navTitle}}"
+    theme-class="{{themeClass}}"
+    show-back="{{true}}"
+    bindback="onBack"
+  />
 
   <scroll-view scroll-y="{{true}}" class="settings-scroll">
     <view class="settings-content">
@@ -355,28 +360,6 @@
           <text class="settings-hint ts-muted">{{tr.thresholdsHint}}</text>
         </block>
         <text wx:else class="settings-empty ts-muted">{{tr.thresholdsEmpty}}</text>
-      </view>
-
-      <!-- Tappable nav row keeps a card surface for affordance — the
-           chevron + tap area need to feel pressable. -->
-      <view class="ts-card settings-nav-card" bindtap="onNavigateToScience">
-        <view class="settings-nav-row">
-          <view>
-            <text class="settings-nav-title">{{tr.trainingScience}}</text>
-            <text class="settings-nav-sub ts-muted">{{tr.scienceSubtitle}}</text>
-          </view>
-          <text class="settings-nav-chevron ts-muted">›</text>
-        </view>
-      </view>
-
-      <view class="ts-card settings-nav-card" bindtap="onNavigateToLabs">
-        <view class="settings-nav-row">
-          <view>
-            <text class="settings-nav-title">{{tr.labs}}</text>
-            <text class="settings-nav-sub ts-muted">{{tr.labsSubtitle}}</text>
-          </view>
-          <text class="settings-nav-chevron ts-muted">›</text>
-        </view>
       </view>
 
       <view class="ts-section">
