+++ miniapp/utils/i18n-catalog.ts
@@ -173,6 +173,7 @@ export const I18N_CATALOG: Record<Locale, Record<string, string>> = {
     "An earlier automatic change was superseded": "An earlier automatic change was superseded",
     "An external planner overlaps the Praxys plan. Use one planner at a time to avoid duplicate or conflicting sessions.": "An external planner overlaps the Praxys plan. Use one planner at a time to avoid duplicate or conflicting sessions.",
     "An unexpected error occurred.": "An unexpected error occurred.",
+    "Analysis": "Analysis",
     "Analysis needs a manual retry": "Analysis needs a manual retry",
     "Analysis request saved": "Analysis request saved",
     "Analysis worker is running": "Analysis worker is running",
@@ -620,10 +621,10 @@ export const I18N_CATALOG: Record<Locale, Record<string, string>> = {
     "Failed to generate invitation code.": "Failed to generate invitation code.",
     "Failed to load": "Failed to load",
     "Failed to load activities": "Failed to load activities",
+    "Failed to load analysis": "Failed to load analysis",
     "Failed to load goal data": "Failed to load goal data",
     "Failed to load science data.": "Failed to load science data.",
     "Failed to load settings": "Failed to load settings",
-    "Failed to load training data": "Failed to load training data",
     "Failed to load training plan": "Failed to load training plan",
     "Failed to revoke invitation code.": "Failed to revoke invitation code.",
     "Failed to save goal": "Failed to save goal",
@@ -1454,6 +1455,7 @@ export const I18N_CATALOG: Record<Locale, Record<string, string>> = {
     "Review delivery queue": "Review delivery queue",
     "Review first issue": "Review first issue",
     "Review goal": "Review goal",
+    "Review load, volume, heat adaptation, and Praxys Coach.": "Review load, volume, heat adaptation, and Praxys Coach.",
     "Review optional test": "Review optional test",
     "Review plan outcomes": "Review plan outcomes",
     "Review purpose and expiry": "Review purpose and expiry",
@@ -1598,6 +1600,7 @@ export const I18N_CATALOG: Record<Locale, Record<string, string>> = {
     "Standard": "Standard",
     "Start sync": "Start sync",
     "Start with operator action, then use health and usage context to choose the next workflow.": "Start with operator action, then use health and usage context to choose the next workflow.",
+    "Start, review, and adjust the plan Praxys manages for you.": "Start, review, and adjust the plan Praxys manages for you.",
     "Started": "Started",
     "Starts": "Starts",
     "Status": "Status",
@@ -2215,6 +2218,7 @@ export const I18N_CATALOG: Record<Locale, Record<string, string>> = {
     "An earlier automatic change was superseded": "此前的自动调整已被后续更改覆盖",
     "An external planner overlaps the Praxys plan. Use one planner at a time to avoid duplicate or conflicting sessions.": "其他计划源与 Praxys 计划有重叠。建议一次只使用一个计划源，避免重复或冲突训练。",
     "An unexpected error occurred.": "发生了意外错误。",
+    "Analysis": "训练分析",
     "Analysis needs a manual retry": "分析需要手动重试",
     "Analysis request saved": "分析请求已保存",
     "Analysis worker is running": "分析任务正在运行",
@@ -2662,10 +2666,10 @@ export const I18N_CATALOG: Record<Locale, Record<string, string>> = {
     "Failed to generate invitation code.": "生成邀请码失败。",
     "Failed to load": "加载失败",
     "Failed to load activities": "加载活动失败",
+    "Failed to load analysis": "分析加载失败",
     "Failed to load goal data": "目标数据加载失败",
     "Failed to load science data.": "加载科学数据失败。",
     "Failed to load settings": "加载设置失败",
-    "Failed to load training data": "训练数据加载失败",
     "Failed to load training plan": "训练计划加载失败",
     "Failed to revoke invitation code.": "撤销邀请码失败。",
     "Failed to save goal": "保存目标失败",
@@ -3496,6 +3500,7 @@ export const I18N_CATALOG: Record<Locale, Record<string, string>> = {
     "Review delivery queue": "查看下发队列",
     "Review first issue": "查看第一个问题",
     "Review goal": "复核目标",
+    "Review load, volume, heat adaptation, and Praxys Coach.": "查看负荷、训练量、热适应和 Praxys Coach 分析。",
     "Review optional test": "查看可选测试",
     "Review plan outcomes": "回顾计划结果",
     "Review purpose and expiry": "确认用途和保留期限",
@@ -3640,6 +3645,7 @@ export const I18N_CATALOG: Record<Locale, Record<string, string>> = {
     "Standard": "标准",
     "Start sync": "开始同步",
     "Start with operator action, then use health and usage context to choose the next workflow.": "先处理运维事项，再结合系统健康和使用情况选择下一步流程。",
+    "Start, review, and adjust the plan Praxys manages for you.": "开始、查看并调整由 Praxys 管理的训练计划。",
     "Started": "开始时间",
     "Starts": "开始日期",
     "Status": "状态",
