+++ tests/test_science_contribution_workflow.py
@@ -138,9 +138,11 @@ def test_science_review_is_layered_and_human_owned() -> None:
         "/web/src/lib/workout-parser.ts",
         "/web/src/pages/Science.tsx",
         "/web/src/pages/Training.tsx",
+        "/web/src/pages/Analysis.tsx",
         "/web/src/types/api.ts",
         "/miniapp/pages/science/",
         "/miniapp/pages/training/",
+        "/miniapp/pages/analysis/",
         "/miniapp/pages/settings/",
         "/miniapp/utils/heat-adaptation.ts",
         "/miniapp/utils/i18n-extra.ts",
@@ -169,8 +171,10 @@ def test_science_codeowners_resolve_representative_product_surfaces() -> None:
         "web/src/contexts/ScienceContext.tsx",
         "web/src/lib/workout-parser.ts",
         "web/src/pages/Training.tsx",
+        "web/src/pages/Analysis.tsx",
         "web/src/types/api.ts",
         "miniapp/pages/training/index.wxml",
+        "miniapp/pages/analysis/index.wxml",
         "miniapp/pages/settings/index.ts",
         "miniapp/utils/heat-adaptation.ts",
         "miniapp/utils/i18n-extra.ts",
