+++ tests/test_training_miniapp_detail.py
@@ -1,15 +1,15 @@
-"""Regression coverage for Training metric-sheet selection races."""
+"""Regression coverage for Analysis metric-sheet selection races."""
 
 from pathlib import Path
 
 
 ROOT = Path(__file__).resolve().parent.parent
-TRAINING_PAGE = ROOT / "miniapp" / "pages" / "training" / "index.ts"
+ANALYSIS_PAGE = ROOT / "miniapp" / "pages" / "analysis" / "index.ts"
 
 
 def test_metric_sheet_selection_survives_concurrent_refetch() -> None:
     """Open and close intents must override stale rendered page data."""
-    source = TRAINING_PAGE.read_text(encoding="utf-8")
+    source = ANALYSIS_PAGE.read_text(encoding="utf-8")
 
     assert "pageState._activeMetric = metric;" in source
     assert "pageState._activeMetric = '';" in source
